chore: refactor api.py in several files
