+++ app/__init__.py
@@ -1,0 +1,6 @@
+"""Speaking Meeting Bot API package."""
+
+from app.main import create_app
+
+# Create app instance using the factory function
+app = create_app()
